backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0
motor==3.3.2
google-generativeai==0.8.4
python-dotenv==1.0.0
pdf2image==1.17.0
pytesseract==0.3.10
pillow==10.1.0
python-magic==0.4.27